+++ RailroadProperty.h
@@ -17,7 +17,7 @@ class RailroadProperty final : public Property {
         virtual string getType() const override;
         virtual void act(Player* player) override;
         virtual vector<int> getPriceTable() const override;
-        virtual int getNumHouses() const override;
+        virtual unsigned short int getNumHouses() const override;
 };
 
 #endif
